+++ src/components/browse/browse-pagination.tsx
@@ -8,13 +8,6 @@ import {
 import { PaginationProps } from '@/views/browse-view'
 
 export const PaginationNumbers = ({ currentPage, onPageChange, maxPages }: PaginationProps) => {
-  const handlePageChange = (newPage: number) => {
-    if (newPage >= 1 && newPage <= maxPages) {
-      onPageChange(newPage)
-      window.scrollTo(0, 0)
-    }
-  }
-
   if (maxPages === undefined) {
     return
   }
@@ -24,7 +17,7 @@ export const PaginationNumbers = ({ currentPage, onPageChange, maxPages }: Pagin
       {/* Show previous button if we are not on page 1 (i.e. there is a page to go back to) */}
       {currentPage > 1 && (
         <PaginationItem>
-          <PaginationPrevious onClick={() => handlePageChange(currentPage - 1)} />
+          <PaginationPrevious onClick={() => onPageChange(currentPage - 1)} />
         </PaginationItem>
       )}
 
@@ -34,7 +27,7 @@ export const PaginationNumbers = ({ currentPage, onPageChange, maxPages }: Pagin
 
       {currentPage < maxPages && (
         <PaginationItem>
-          <PaginationNext onClick={() => handlePageChange(currentPage + 1)} />
+          <PaginationNext onClick={() => onPageChange(currentPage + 1)} />
         </PaginationItem>
       )}
     </PaginationContent>
